requirements.txt
click>=8.1.7
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
greenlet>=2.0.2
itsdangerous>=2.2.0
Jinja2>=3.1.4
MarkupSafe>=2.1.1
SQLAlchemy>=2.0.36
Werkzeug>=3.1.3
Flask-Migrate>=4.0.7
Flask-MySQLdb>=2.0.0
marshmallow>=3.23.0
marshmallow-sqlalchemy>=0.26.1
mysql==0.0.3
PyMySQL>=1.1.1

flask
python-dotenv
